agent/requirements.txt
langgraph
langchain
langchain-anthropic
langchain-openai
langchain-community
playwright
pydantic
maturin
python-dotenv
pillow
